requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
Django==5.2.5
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-extra-fields==3.7.0
drf-yasg==1.21.10
environs==14.3.0
filetype==1.2.0
frozenlist==1.7.0
idna==3.10
inflection==0.5.1
marshmallow==4.0.0
multidict==6.6.4
packaging==25.0
propcache==0.3.2
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
yarl==1.20.1
